Lower case tag data & comment about split intervals don't work
Added code to change boolean data getting written to tags to be lower case to match what OSMAnd does.

Found out that the split_type and Split_interval tags don't seem to have an effect once the track GPX file is loaded into OSMAnd.  If I activate the split feature in OSMAnd and then export the GPX it's the same tags.  It appears that OSMAnd isn't using the values from the GPX file.  Arrows and ends both work depending on their true/false values so something is different about split.